odhcpd: add option to use absolute timestamps
Until now it is not possible to give absolute timestamps in odhcpd.
This means that on every new RA or request, the timestamp is renewed.
Further, the valid and preferred lifetimes are not synced between all
devices.

There are several usecases when it is needed to have absolute timestamp
that needed to be synced across all devices, e.g. your ISP delegates
you a prefix for some certain time, or you want to change to another
prefix.

The purpose of having this as a absolute timestamp is to make it easier
to track. An example configuration is

  option absolute_lifetime '1'
  option valid_lifetime '05 Jan 2021 23:00:00'
  option preferred_lifetime '05 Jan 2021 23:00:00'

If the valid_lifetime is in the past, the preferred lifetime and valid
lifetime are set to 1 minute.

Signed-off-by: Nick Hainke <vincent@systemli.org>